Update tag-resources.md
The page displays a list of resource types that are limited to 15 tags. That list is not complete/correct for Azure DNS. It should include the following resource types for Azure DNS:

- Microsoft.Network/dnsZones
- Microsoft.Network/dnsZones/A
- Microsoft.Network/privateDnsZones
- Microsoft.Network/privateDnsZones/virtualNetworkLinks

I have dealt with this in many production environments. I must say I never had an issue with A-records. But the DNS Zones and Private DNS Zones, including its Virtual Network Links are certainly limited to 15 tags.

Many customers use Azure Policy to govern tags. I use a resource selector on policy assignments to exclude tags from these resources.  It is even better to provide de exact resource type in a format like "Microsoft.Dns/...". In my case that would be the following:

- Microsoft.Automation/automationAccounts
- Microsoft.Cdn/profiles
- Microsoft.Network/dnsZones
- Microsoft.Network/dnsZones/A
- Microsoft.Network/privateDnsZones
- Microsoft.Network/privateDnsZones/virtualNetworkLinks
- Microsoft.OperationalInsights/workspaces/savedSearches